requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-cloud-vision==3.4.1
google-cloud-speech==2.21.0
python-dotenv==1.0.0
sqlalchemy
authlib
itsdangerous
httpx
